Add files via upload
A test of homogeneity is used to compare categorical variables and their distributions. The idea of homogeneity is derived from the null hypothesis, which states that we have equal cell probabilities, and as a result our data would come from the same multinomial distribution. Our question of interest is: do the two (or more) multinomial distributions differ? The test will be derived using the likelihood ratio test and the decision of homogeneity will be based on the resulting comparison of the test statistic and the corresponding Chi-Square critical value.